core+helpers: relocate toPaymentInfo to @x402r/helpers (PR 4)
Move `toPaymentInfo` and `ToPaymentInfoReturnType` from
`@x402r/core/payment/serialization` to `@x402r/helpers`. Drop `@x402r/evm`
peerDep+devDep from `@x402r/core` — nothing else in core referenced it.

Add `viem` to `@x402r/helpers` deps (the relocated function imports
`hexToBigInt`). Inline `TEST_ADDRESSES` in the moved test since helpers
has no shared fixtures file.

History preserved via `git mv` on both src and test.
