+++ packages/core/src/index.ts
@@ -359,10 +359,6 @@ export {
   type SignReceiveAuthorizationReturnType,
   signReceiveAuthorization,
 } from './payment/index.js'
-export {
-  type ToPaymentInfoReturnType,
-  toPaymentInfo,
-} from './payment/serialization.js'
 export { validatePaymentInfo } from './payment/validation.js'
 // ---------------------------------------------------------------------------
 // Types
