+++ packages/core/src/payment/index.ts
@@ -10,8 +10,4 @@ export {
   computePaymentInfoHash,
   PAYMENT_INFO_TYPEHASH,
 } from './hashing.js'
-export {
-  type ToPaymentInfoReturnType,
-  toPaymentInfo,
-} from './serialization.js'
 export { validatePaymentInfo } from './validation.js'
