+++ packages/helpers/src/index.ts
@@ -28,5 +28,7 @@ export {
 } from '@x402r/evm'
 export type { ForwardToArbiterOptions } from './forward-to-arbiter.js'
 export { forwardToArbiter } from './forward-to-arbiter.js'
+export type { ToPaymentInfoReturnType } from './serialization.js'
+export { toPaymentInfo } from './serialization.js'
 export type { X402rDefaultsInput } from './x402r-defaults.js'
 export { x402rDefaults } from './x402r-defaults.js'
